feat: add setupscripts and remove hooks
The new operator supports only scripts and hooks are deprecated. We had
several issues and attempts to make it work with hooks but they are
feeling kind of hacky. So we provide a possiblity to overwrite the
command completely and let the user decide what to do, if a setup is
crashing or when it should crash.