refactore: fix issue with lookup
We remove the lookup function here. We implemented this for argoCD but
unfortnatly this is not working with argocd. Thats why you should set it
ourself. This doesn't change the behaviour so for helm upgrade so there
should be no issues.
Fourthermore this helps also with helm template and helm lint